+++ inst/doc/AOA-tutorial.R
@@ -1,7 +1,7 @@
-## ----setup, echo=FALSE---------------------------------------------------
+## ----setup, echo=FALSE--------------------------------------------------------
 knitr::opts_chunk$set(fig.width = 8.83)
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
+## ---- message = FALSE, warning=FALSE------------------------------------------
 library(CAST)
 library(virtualspecies)
 library(caret)
@@ -12,20 +12,20 @@ library(viridis)
 library(latticeExtra)
 library(gridExtra)
 
-## ----message = FALSE,include=FALSE, warning=FALSE------------------------
+## ----message = FALSE,include=FALSE, warning=FALSE-----------------------------
 RMSE = function(a, b){
     sqrt(mean((a - b)^2,na.rm=T))
 }
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
+## ---- message = FALSE, warning=FALSE------------------------------------------
 predictors <- stack(system.file("extdata","bioclim.grd",package="CAST"))
 spplot(stretch(predictors,0,1),col.regions=viridis(100))
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 response <- generateSpFromPCA(predictors,
                               means = c(3,1),sds = c(2,2), plot=F)$suitab.raster
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 mask <- predictors[[1]]
 values(mask)[!is.na(values(mask))] <- 1
 mask <- rasterToPolygons(mask)
@@ -34,12 +34,12 @@ samplepoints <- spsample(mask,20,"random")
 spplot(response,col.regions=viridis(100),
             sp.layout=list("sp.points", samplepoints, col = "red", first = FALSE, cex=2))
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 trainDat <- extract(predictors,samplepoints,df=TRUE)
 trainDat$response <- extract (response,samplepoints)
 trainDat <- trainDat[complete.cases(trainDat),]
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 set.seed(10)
 model <- train(trainDat[,names(predictors)],
                trainDat$response,
@@ -49,25 +49,25 @@ model <- train(trainDat[,names(predictors)],
 print(model)
 
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 plot(varImp(model,scale = F),col="black")
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 prediction <- predict(predictors,model)
 truediff <- abs(prediction-response)
 spplot(stack(prediction,response),main=c("prediction","reference"))
 
-## ----message = FALSE, warning=FALSE--------------------------------------
-AOA <- aoa(trainDat,predictors, variables = names(predictors),model=model)
+## ----message = FALSE, warning=FALSE-------------------------------------------
+AOA <- aoa(predictors, model)
 attributes(AOA)$aoa_stats
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 grid.arrange(
   spplot(truediff,col.regions=viridis(100),main="true prediction error"),
   spplot(AOA$DI,col.regions=viridis(100),main="DI"),
-  spplot(prediction, col.regions=viridis(100),main="prediction for AOA")+ spplot(AOA$AOA,col.regions=c("transparent","grey")), ncol=3)
+  spplot(prediction, col.regions=viridis(100),main="prediction for AOA")+ spplot(AOA$AOA,col.regions=c("grey","transparent")), ncol=3)
 
-## ----clusteredpoints,message = FALSE, include=FALSE----------------------
+## ----clusteredpoints,message = FALSE, include=FALSE---------------------------
 #For a clustered sesign:
 csample <- function(x,n,nclusters,maxdist,seed){
   set.seed(seed)
@@ -86,20 +86,20 @@ csample <- function(x,n,nclusters,maxdist,seed){
   return(result)
 }
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 
 samplepoints <- csample(mask,75,15,maxdist=0.20,seed=15)
 spplot(response,col.regions=viridis(100),
             sp.layout=list("sp.points", samplepoints, col = "red", first = FALSE, cex=2))
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 
 trainDat <- extract(predictors,samplepoints,df=TRUE)
 trainDat$response <- extract (response,samplepoints)
 trainDat <- merge(trainDat,samplepoints,by.x="ID",by.y="ID")
 trainDat <- trainDat[complete.cases(trainDat),]
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 set.seed(10)
 model_random <- train(trainDat[,names(predictors)],
                trainDat$response,
@@ -109,7 +109,7 @@ model_random <- train(trainDat[,names(predictors)],
 prediction_random <- predict(predictors,model_random)
 print(model_random)
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 folds <- CreateSpacetimeFolds(trainDat, spacevar="clstrID",k=10)
 set.seed(15)
 model <- train(trainDat[,names(predictors)],
@@ -122,12 +122,12 @@ model <- train(trainDat[,names(predictors)],
   
 prediction <- predict(predictors,model)
 
-## ----message = FALSE, warning=FALSE--------------------------------------
-AOA_spatial <- aoa(trainDat,predictors, variables = names(predictors),model=model)
+## ----message = FALSE, warning=FALSE-------------------------------------------
+AOA_spatial <- aoa(predictors, model)
 
-AOA_random <- aoa(trainDat,predictors, variables = names(predictors),model=model_random)
+AOA_random <- aoa(predictors, model_random)
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 grid.arrange(spplot(AOA_spatial$DI,col.regions=viridis(100),main="DI"),
   spplot(prediction, col.regions=viridis(100),main="prediction for AOA \n(spatial CV error applies)")+
          spplot(AOA_spatial$AOA,col.regions=c("grey","transparent")),
@@ -135,7 +135,7 @@ grid.arrange(spplot(AOA_spatial$DI,col.regions=viridis(100),main="DI"),
          spplot(AOA_random$AOA,col.regions=c("grey","transparent")),
 ncol=3)
 
-## ----message = FALSE, warning=FALSE--------------------------------------
+## ----message = FALSE, warning=FALSE-------------------------------------------
 ###for the spatial CV:
 RMSE(values(prediction)[values(AOA_spatial$AOA)==1],values(response)[values(AOA_spatial$AOA)==1])
 RMSE(values(prediction)[values(AOA_spatial$AOA)==0],values(response)[values(AOA_spatial$AOA)==1])
@@ -146,7 +146,7 @@ RMSE(values(prediction_random)[values(AOA_random$AOA)==1],values(response)[value
 RMSE(values(prediction_random)[values(AOA_random$AOA)==0],values(response)[values(AOA_random$AOA)==1])
 model_random$results
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
+## ---- message = FALSE, warning=FALSE------------------------------------------
 dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
 # calculate average of VW for each sampling site:
 dat <- aggregate(dat[,c("VW","Easting","Northing")],by=list(as.character(dat$SOURCEID)),mean)
@@ -162,7 +162,7 @@ trainDat <- merge(trainDat,pts,by.x="ID",by.y="ID")
 # The final training dataset with potential predictors and VW:
 head(trainDat)
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
+## ---- message = FALSE, warning=FALSE------------------------------------------
 predictors <- c("DEM","NDRE.Sd","TWI","Bt")
 response <- "VW"
 
@@ -171,19 +171,19 @@ model <- train(trainDat[,predictors],trainDat[,response],
                trControl=trainControl(method="LOOCV"))
 model
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
+## ---- message = FALSE, warning=FALSE------------------------------------------
 #Predictors:
 spplot(stretch(studyArea[[predictors]]))
 
 #prediction:
 prediction <- predict(studyArea,model)
 
-## ---- message = FALSE, warning=FALSE-------------------------------------
-AOA <- aoa(trainDat,studyArea,variables=predictors,model=model)
+## ---- message = FALSE, warning=FALSE------------------------------------------
+AOA <- aoa(studyArea,model)
 
 #### Plot results:
 grid.arrange(spplot(AOA$DI,col.regions=viridis(100),main="DI with sampling locations (red)")+
                spplot(as_Spatial(pts),zcol="ID",col.regions="red"),
-  spplot(prediction, col.regions=viridis(100),main="prediction for AOA \n(LOOCV error applies)")+ spplot(AOA$AOA,col.regions=c("transparent","grey")),ncol=2)
+  spplot(prediction, col.regions=viridis(100),main="prediction for AOA \n(LOOCV error applies)")+ spplot(AOA$AOA,col.regions=c("grey","transparent")),ncol=2)